+++ rotkehlchen/chain/arbitrum_one/modules/aave/v3/decoder.py
@@ -26,6 +26,7 @@ class Aavev3Decoder(Aavev3LikeCommonDecoder):
             native_gateways=(
                 string_to_evm_address('0x5760E34c4003752329bC77790B1De44C2799F8C3'),
                 string_to_evm_address('0xecD4bd3121F9FD604ffaC631bF6d41ec12f1fafb'),
+                string_to_evm_address('0x5283BEcEd7ADF6D003225C13896E536f2D4264FF'),
             ),
             treasury=string_to_evm_address('0x053D55f9B5AF8694c503EB288a1B7E552f590710'),
             incentives=string_to_evm_address('0x929EC64c34a17401F460460D4B9390518E5B473e'),